report successful startup to the platform
For proper CCD password management it is necessary to be able to alert
the rest of the system when the actual restart happens. A good
indication of a restart is successful processing of the Starup command
of the SU_RESET type.

CQ-DEPEND=CL:804141
BRANCH=cr50
BUG=b:67007578
TEST=verified that the callback is indeed called on appropriate starups.

Change-Id: I984fb971410ff29454250c40e44a7ca7710eaf1c
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/803741
Reviewed-by: Andrey Pronin <apronin@chromium.org>
(cherry picked from commit ae8d044cd15f49f56de4db84da7203a0ac2ff783)
Reviewed-on: https://chromium-review.googlesource.com/820606