Implement _plat__ShallSurviveOwnerClear for non-cr50 use
libtpm2 is used outside of cr50 for fuzzing and tpm2_simulator.
Before this CL, _plat__ShallSurviveOwnerClear was implemented in
platform/ec only. So, attempts to link with libtpm2 in other
contexts led to undefined reference to `_plat__ShallSurviveOwnerClear'.

Instead of implementing separately for each situation, implement the
default NOP _plat__ShallSurviveOwnerClear in NVMem.c, which as a part of
HOST_SOURCES is included in libtpm2 when built in non-embedded context.

BUG=oss-fuzz:9793
TEST=in third_prty/tpm2:
     docker build -t ossfuzz/tpm2 -f fuzz/Dockerfile .
     docker run -ti --rm -v $(pwd):/src/tpm2 -v /tmp/fuzzers:/out \
     ossfuzz/tpm2

Change-Id: I0cc3224d7568d40fc031b0901705631fd6618143
Signed-off-by: Andrey Pronin <apronin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1171514
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>