NV: add function to selectively wipe out NVMEM cache
The function receives a pointer to a two element array specifying the
inclusive range of NV indices to preserve, the rest of the evictable
NVMEM cache objects need to be corrupted.

The function finds offsets of the objects, and overwrites objects'
contents with arbitrary data from the text segment.

BRANCH=cr50, cr50mp
BUG=b:118504817
TEST=checked along with corresponding modifications of Cr50 and AP
     firmware. Using a test function verified that non-excluded
     objects get overwritten.

Change-Id: Ib753b892f36c800a9cfe9b8a297708b760173ee0
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1362161
Reviewed-by: Namyoon Woo <namyoon@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>