NV: provide a way to find location of reserved objects
The new NVMEM flash storage scheme is operating on separate TPM
objects as opposed to saving the entire NVMEM cache space each time
anything changes. That is the new storage scheme implementation needs
information about cache locations and sizes of reserved NVMEM objects.

Note that on top of the 'legit' reserved objects in the NVMEM cache
there are also RAM index space and the maxCount. These two fields are
supposed to be always preserved in the NVMEM, and they are located at
fixed offsets in the cache. This is why they are being added to the
set of TPM reserved objects from the caller point of view, accessed at
indices NV_RESERVE_LAST and NV_RESERVE_LAST + 1 respectively.

The RAM index space is of variable size, its actual present size is
returned to the caller.

BRANCH=cr50, cr50-mp
BUG=b:69907320
TEST=with the rest of TPM NVMEM platform changes verified that ec and
      TCG tests pass.

Change-Id: I8bfe33f2dfa7b276848b7a08a78f2b7f05a86637
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1447091
Reviewed-by: Andrey Pronin <apronin@chromium.org>