Reduce TPM buffer size
We are getting tight on SRAM, and the platform receive/transmit buffer
is 2K in size anyways, no point in keeping 4K in the tpm library.
Ideally we should try eliminating this duplication and have one buffer
with zero copy, but this is for another day.

BRANCH=cr50, cr50-mp
BUG=b:65253310
TEST=TCG test suite passes successfully

Change-Id: I9c64ae1759aa1a4780438d884b3d7f52275fe66e
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1467622
Reviewed-by: Andrey Pronin <apronin@chromium.org>