generator: update generator sources to better match generated files
It is not entirely clear why, but the generated TPM library files now
differ from what was checked in earlier.

The discrepancies include comment alignment, year in the copyright
header, wrapping of the short structures, and manual modifications
tpm_generated.c and tpm_types.h.

Wrapping of short structures must be a side effect of changes to
clang-format, but the rest of the discrepancies can be fixed by
modifying the generator.

A patch file containing Chrome OS specific changes is also being
added.

BRANCH=cr50, cr50mp
BUG=b:65253310

TEST=verified that generated source code (uploaded as the next patch)
     is identical to the checked it code modulo structure inlining.

Change-Id: I0e585574abaec676ca7853080b5c7284bf02d0a9
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1413093
Reviewed-by: Andrey Pronin <apronin@chromium.org>