Makefile: provide a target to build object files with renamed .bss
This would be especially useful when link time optimization is
enabled, as in that case the linker does not operate properly on
archives, separate object files are required.

Using a non-standard name for the .bss segment makes it simpler to
properly place this library's uninitialized data in the executable
image.

Also, prevent multiple execution of string substitution function when
deriving lists of object and dependency file names from source file
names.

BRANCH=cr50, cr50-mp
BUG=b:65253310
TEST=Cr50 image builds fine with object files produced when using the
     new make target during build

Change-Id: Id1d654a2e9d8a3b5e361e85dac354b37bf1511d3
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1457398
Reviewed-by: Randall Spangler <rspangler@chromium.org>