Makefile: prevent LTO mode compilation
Due to the fact that LTO compiled objects do not provide a clear
visibility into the .bss section, make sure that even if the caller
enabled LTO mode compilation, the modules are compiled without it.

If this is not enforced, linker fails put TPM2 library .bss segment
into the area which is clreaed on TPM restarts.

BRANCH=cr50, cr50-mp
BUG=b:65253310

TEST=verified that produced .elf Cr50 images have expected .bss
     section layout.

Change-Id: I831b472c5f751f4280ddf9faed73a008f03d03a9
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1459742
Reviewed-by: Randall Spangler <rspangler@chromium.org>