make: Only prevent LTO when building for EMBEDDED_MODE (take two)
This line to prevent LTO was added in CL:1459742. It is causing some
trouble for us in crosvm where we would like to use libtpm2 as a library
and still build all of crosvm with LTO.

It seems like the change to prevent LTO was motivated by cr50's use case
for tpm2, for reasons that do not apply to crosvm's use of tpm2. (Please
correct me if that is wrong and you believe anything using tpm2 would
need to build without LTO in order to use the library correctly. For
what it's worth, it seems to have been working fine until that CL.)

This CL changes the Makefile to only prevent LTO in the case that
EMBEDDED_MODE is set, as set by:
https://chromium.googlesource.com/chromiumos/platform/ec/+/d7aab8081efcd6187ae1704baa989cf973119380/board/cr50/build.mk#107

BUG=chromium:911799
TEST=make, make EMBEDDED_MODE=1

Reviewed-on: https://chromium-review.googlesource.com/1548712
Commit-Ready: David Tolnay <dtolnay@chromium.org>
Tested-by: David Tolnay <dtolnay@chromium.org>
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
(cherry picked from commit ea24233c2bb4f7e4a31aec478d45122fda79739d)

Change-Id: I5ddad466bcece20dcbcb9baeeb7b0e2728259595
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/tpm2/+/1644183
Tested-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Commit-Queue: Vadim Bendebury <vbendeb@chromium.org>