step back to 0.3.26 state
The MP branch got ahead of itself, we need to publish an intermediate
release before 5.4. Let's bring the branch back to 0.3.26 state so
changes necessary for the intermediate release can be made cleanly.
Once the release is cut, this patch will be reverted to bring the
branch back to the top.

BRANCH=none
BUG=none
TEST='git diff beddeb2' returns no output.

Change-Id: I8b2cad19f6fdfd271e9fafb17ec70e8d3d956a7d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/tpm2/+/2189880
Tested-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-by: Mary Ruthven <mruthven@chromium.org>
Commit-Queue: Vadim Bendebury <vbendeb@chromium.org>